requirements.txt
Flask>=2.3.0
SpeechRecognition>=3.8.1
pyttsx3>=2.90
requests>=2.31.0
groq>=0.4.1
python-dotenv>=1.0.0

